+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ThrownEggMixin.java
@@ -1,15 +1,15 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.iguanatweaksreborn.module.farming.livestock.Livestock;
 import insane96mcp.insanelib.base.Feature;
 import net.minecraft.world.entity.projectile.ThrownEgg;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
+import org.spongepowered.asm.mixin.injection.At;
 
 @Mixin(ThrownEgg.class)
 public class ThrownEggMixin {
-    @ModifyConstant(method = "onHit", constant = @Constant(intValue = 8))
+    @ModifyExpressionValue(method = "onHit", at = @At(value = "CONSTANT", args = "intValue=8"))
     public int chanceForChicken(int chance) {
         return !Feature.isEnabled(Livestock.class) ? chance : Livestock.chickenFromEggChance;
     }